# proxy/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI Cost Control Tower — Ingestion API
  version: "1.0.0"
  license:
    name: Internal use only — not for external distribution
  description: |
    ## Getting started

    This API lets an internal HCLTech application write its own AI usage and
    outcome events into the same Cost Control Tower tables Product Studio
    already uses, so it can be traced, costed, and governed the same way.

    **Registration is not self-service.** Contact the platform operator to
    register your app and receive a credential — registration and credential
    issuance are direct SQL/RPC, run manually, not exposed over HTTP
    (see the spec's Section 3/4 for why).

    **Credential handling.** Your API key is shown exactly once, at issuance
    time. Move it into your own team's secrets manager immediately — it is
    never stored in plaintext or recoverable after that point. Every request
    authenticates with `Authorization: Bearer <api_key>`; `company_id` and
    `app_id` are resolved from the credential server-side and are never
    accepted as request fields.

    ```bash
    curl -X POST https://<ingestion-host>/v1/usage-events \
      -H "Authorization: Bearer <your_api_key>" \
      -H "Content-Type: application/json" \
      -d '{
        "client_call_id": "a1b2c3d4-...",
        "user_role_at_call": "engineer",
        "caller": "my-app-feature-x",
        "requested_model": "gpt-4o",
        "status": "success",
...
      }'
    ```

...

...
      shape.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/traces:
    post:
      operationId: createTrace
      tags: [traces]
      summary: Create a trace explicitly, ahead of your first usage event
      description: |
        Not the only way a trace comes into existence — `POST /v1/usage-events`
        creates one implicitly via `client_trace_id` if you haven't called
        this first. Use this when you want `product_id`/`session_id`/
        `outcome_id` set at creation time rather than picked up from your
        first usage event. Idempotent on `client_trace_id`: a resubmission
        with the same `agent_name`/`session_id` is a no-op; a mismatch is a
        409 (a genuine identity conflict on your own idempotency key, not
        something to silently paper over).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_name, client_trace_id]
              properties:
                agent_name: { type: string }
                client_trace_id: { type: string, description: "Your own UUID, generated once per conversation/agent run" }
                session_id: { type: string, nullable: true, description: "Intentionally not required — no session concept is forced on you" }
                product_id: { type: string, format: uuid, nullable: true }
                outcome_id: { type: string, format: uuid, nullable: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  trace_id: { type: string, format: uuid }
                  deduplicated: { type: boolean }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: client_trace_id already registered with a different agent_name or session_id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorBody' }
  /v1/traces/{id}:
    get:
      operationId: getTrace
      tags: [traces]
      summary: Read one trace
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TraceRecord' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: completeTrace
      tags: [traces]
      summary: Mark a trace completed
      description: Sets `completed_at` only. Ownership-checked by `(company_id, app_id)`.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  trace_id: { type: string, format: uuid }
                  completed_at: { type: string, format: date-time }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/traces/{id}/spans:
    get:
      operationId: listTraceSpans
      tags: [traces]
      summary: "The \"why did this run cost 4x average\" query"
      description: |
        Every span in the trace, in `sequence_order`. An `llm_call` span
        embeds its `mt_ai_usage_events` row's cost/token/model/duration
        fields; a `tool_call` span has none (its own `duration_ms` is stored
        directly on the span instead, since it may involve no LLM call at all).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  spans:
                    type: array
                    items: { $ref: '#/components/schemas/SpanRecord' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The API key issued to your app, as a Bearer token.
  responses:
    BadRequest:
      description: Malformed payload
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    Unauthorized:
      description: Missing, invalid, or inactive credential
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    NotFound:
      description: Referenced resource doesn't exist (or doesn't belong to the caller)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            type: { type: string }
            message: { type: string }
...
    TraceRecord:
      type: object
      properties:
        trace_id: { type: string, format: uuid }
        agent_name: { type: string }
        session_id: { type: string, nullable: true }
        product_id: { type: string, format: uuid, nullable: true }
        outcome_id: { type: string, format: uuid, nullable: true }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
    SpanRecord:
      type: object
      properties:
        span_id: { type: string, format: uuid }
        parent_span_id: { type: string, format: uuid, nullable: true }
        span_type: { type: string, enum: [llm_call, tool_call] }
        tool_name: { type: string, nullable: true, description: "Set only for a tool_call span" }
        sequence_order: { type: integer }
        attempt_number: { type: integer, nullable: true }
        status: { type: string, enum: [success, error, timeout] }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        duration_ms:
          type: integer
          nullable: true
          description: "Only populated directly for a tool_call span — an llm_call span's duration comes from its embedded usage-event object instead."
        mt_ai_usage_events:
          type: object
          nullable: true
          description: "Present only for an llm_call span; null for tool_call."
          properties:
            requested_model: { type: string }
            response_model: { type: string, nullable: true }
            provider: { type: string }
            input_tokens: { type: integer, nullable: true }
            output_tokens: { type: integer, nullable: true }
            duration_ms: { type: integer, nullable: true }